move news carousel to home section
